Now uploaded complete dataset (bug in writing csv made me miss most of the initial database). Also changed format to commas from semicolons and year to year/01/01 for broader compatibility.
Was missing a large part of the dataset in the write-out to a csv file. Now fixed and added.

Excel works well with semicolon separated csv files, but most of the internet is not impressed. Changed to commas, and using double quotes in names and notes to escape commas. Excel can

Current format hopefully plays nice with both github and kepler.gl