Manually checking Noordegraaf again. Updated the coordinates of Noorderkwartier, and added four entries that were missing
Removed Scheveningen 1568 - although that is listed in Bijlage 2, it is not listed in Bijlage 1 which contains references to the sources, and the only source mentioned for Scheveningen in the book is dealing with the period 1600-1800.

Reversely, entries for Haarlem (1555), Enkhuizen (1637) and Scheveningen (1656) are
listed and sourced in Bijlage 1, but are missing from Bijlage 2. Looks like bijlage 2 is just less reliable.